[nvme] detect nvme version

+++ kernel/include/pci.h
@@ -2,5 +2,5 @@
 #include <stdint.h>
 
 void init_pci(void);
-int pci_enable_msi(uint8_t bus, uint8_t device, uint8_t function,
+uintptr_t pci_enable_msi(uint8_t bus, uint8_t device, uint8_t function,
         uint8_t capabilities_list, uint32_t handler_vector);
